+++ apps/client/src/app/pages/profile/public-profile/public-profile.component.ts
@@ -12,6 +12,9 @@ import { GearSet } from '@ffxiv-teamcraft/simulator';
 import { CraftingRotation } from '../../../model/other/crafting-rotation';
 import { CraftingRotationService } from '../../../core/database/crafting-rotation/crafting-rotation.service';
 import { FirestoreListStorage } from '../../../core/database/storage/list/firestore-list-storage';
+import { Apollo } from 'apollo-angular';
+import gql from 'graphql-tag';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-public-profile',
@@ -30,13 +33,16 @@ export class PublicProfileComponent {
 
   public communityRotations$: Observable<CraftingRotation[]>;
 
-  notFound = false;
+  public fishingRankings$: Observable<any>;
 
-  now = Math.floor(Date.now() / 1000);
+  public notFound = false;
+
+  public now = Math.floor(Date.now() / 1000);
 
   constructor(private route: ActivatedRoute, private characterService: CharacterService,
               private listsService: FirestoreListStorage, private authFacade: AuthFacade,
-              private userService: UserService, private craftingRotationsService: CraftingRotationService) {
+              private userService: UserService, private craftingRotationsService: CraftingRotationService,
+              private apollo: Apollo, public translate: TranslateService) {
     const userId$ = this.route.paramMap.pipe(
       map(params => params.get('userId')),
       shareReplay(1)
@@ -66,6 +72,48 @@ export class PublicProfileComponent {
         return this.craftingRotationsService.getUserCommunityRotations(userId);
       }),
       shareReplay(1)
+    );
+    this.fishingRankings$ = this.user$.pipe(
+      switchMap(user => {
+        return this.apollo.query<any>({
+          query: gql`query userRankings {
+            user_rankings:fishingresults(where:{userId: {_eq: "${user.$key}"}, ranking: { rank: {_lte: 3}}, itemId: {_gt: 0}}) {
+                size,
+                date,
+                itemId,
+                baitId,
+                ranking {
+                  rank
+                }
+            }
+          }`,
+          fetchPolicy: 'no-cache'
+        });
+      }),
+      map(result => {
+        const dataDisplay = result.data.user_rankings
+          .reduce((display, row) => {
+            const displayRow = display.find(dRow => {
+              return dRow.itemId === row.itemId;
+            });
+            if (displayRow && displayRow.ranking.rank <= row.ranking.rank) {
+              display = display.filter(dRow => dRow.itemId !== row.itemId);
+            }
+            display.push(row);
+            return display;
+          }, [])
+          .sort((a, b) => {
+            return a.ranking.rank - b.ranking.rank;
+          });
+        return {
+          summary: {
+            1: dataDisplay.filter(row => row.ranking.rank === 1).length,
+            2: dataDisplay.filter(row => row.ranking.rank === 2).length,
+            3: dataDisplay.filter(row => row.ranking.rank === 3).length,
+          },
+          data: dataDisplay
+        };
+      })
     );
   }