+++ apps/client/src/environments/patch-notes.ts
@@ -1,24 +1,13 @@
 export const patchNotes = `### Bug Fixes
 
-* **db:** fixed bait filter in fishing-spot page's bite time graph.
-* **desktop:** fixed merge from different inventory panels counting for autofill.
-* **favorites:** fixed infinite loading for workshops.
-* **gearset:** fixed search input for minimum ilvl not being updated properly.
-* **list:** navigation optimizer now takes panel type into account.
-* **lists:** copying a list now updates it automatically.
-* **lists:** hide archived lists in list picker.
+* **db:** fish rankings are no longer affected by spot filter.
+* **profile:** fixed jobs panel display.
+* **search:** better handling of search filters and overall form display.
+* **search:** first option in history isn't automatically selected anymore.
+* **tutorial-popup:** fixed missing translation.
 
 
 ### Features
 
-* new patch notes popup.
-* **core:** new loading gif.
-* **db:** added ranking in fish details page.
-* **desktop:** new toggle to download update when available, defaults to yes.
-* **gearset:** changed layout in display.
-* **gearset:** changed materia order in materia popup.
-* **gearset:** changed size of the stats tables to match the largest one.
-* **inventory-optimizer:** maximum price filter for vendors optimizer.
-* **patreon:** new patron: Kittie Purry.
-* **recipe-finder:** added +1 button to add only one recipe.
-* added an interactive tour system for various pages.`;
+* **profile:** added a public profile link in place of the removed apkallu falls button.
+* **profile:** added fishing rankings summary in public profile.`;